ADD: support config like whether push down to es
