+++ docs/en/docs/lakehouse/multi-catalog/es.md
@@ -50,16 +50,17 @@ After switching to the ES Catalog, you will be in the `dafault_db`  so you don't
 
 ### Parameter Description
 
-| Parameter         | Required or Not | Default Value | Description                                                  |
-| ----------------- | --------------- | ------------- | ------------------------------------------------------------ |
-| `hosts`           | Yes             |               | ES address, can be one or multiple addresses, or the load balancer address of ES |
-| `user`            | No              | Empty         | ES username                                                  |
-| `password`        | No              | Empty         | Password of the corresponding user                           |
-| `doc_value_scan`  | No              | true          | Whether to obtain value of the target field by ES/Lucene columnar storage |
-| `keyword_sniff`   | No              | true          | Whether to sniff the text.fields in ES based on keyword; If this is set to false, the system will perform matching after tokenization. |
+| Parameter         | Required or Not | Default Value | Description                                                                                                                                       |
+| ----------------- | --------------- | ------------- |---------------------------------------------------------------------------------------------------------------------------------------------------|
+| `hosts`           | Yes             |               | ES address, can be one or multiple addresses, or the load balancer address of ES                                                                  |
+| `user`            | No              | Empty         | ES username                                                                                                                                       |
+| `password`        | No              | Empty         | Password of the corresponding user                                                                                                                |
+| `doc_value_scan`  | No              | true          | Whether to obtain value of the target field by ES/Lucene columnar storage                                                                         |
+| `keyword_sniff`   | No              | true          | Whether to sniff the text.fields in ES based on keyword; If this is set to false, the system will perform matching after tokenization.            |
 | `nodes_discovery` | No              | true          | Whether to enable ES node discovery, set to true by default; set to false in network isolation environments and only connected to specified nodes |
-| `ssl`             | No              | false         | Whether to enable HTTPS access mode for ES, currently follows a "Trust All" method in FE/BE |
-| `mapping_es_id`   | No              | false         | Whether to map the  `_id`  field in the ES index             |
+| `ssl`             | No              | false         | Whether to enable HTTPS access mode for ES, currently follows a "Trust All" method in FE/BE                                                       |
+| `mapping_es_id`   | No              | false         | Whether to map the  `_id`  field in the ES index                                                                                                  |
+| `like_push_down`  | No              | true          | Whether to transform like to wildcard push down to es, this increases the cpu consumption of the es.                                              |
 
 > 1. In terms of authentication, only HTTP Basic authentication is supported and it requires the user to have read privilege for the index and paths including `/_cluster/state/` and `_nodes/http` ; if you have not enabled security authentication for the cluster, you don't need to set the  `user` and `password`.
 >
