+++ fe/fe-core/src/main/java/org/apache/doris/catalog/EsTable.java
@@ -87,9 +87,14 @@ public class EsTable extends Table {
     // would downgrade to extract value from `stored_fields`
     private int maxDocValueFields = DEFAULT_MAX_DOCVALUE_FIELDS;
 
+    // Whether to enable the discovery of es nodes, You can disable it if you are in network isolation
     private boolean nodesDiscovery = Boolean.parseBoolean(EsResource.NODES_DISCOVERY_DEFAULT_VALUE);
 
+    // Whether to use ssl call es, be and fe access through trust
     private boolean httpSslEnabled = Boolean.parseBoolean(EsResource.HTTP_SSL_ENABLED_DEFAULT_VALUE);
+
+    // Whether pushdown like expr, like will trans to wildcard query, consumes too many es cpu resources
+    private boolean likePushDown = Boolean.parseBoolean(EsResource.LIKE_PUSH_DOWN_DEFAULT_VALUE);
 
     // tableContext is used for being convenient to persist some configuration parameters uniformly
     private Map<String, String> tableContext = new HashMap<>();
@@ -171,6 +176,10 @@ public class EsTable extends Table {
             httpSslEnabled = EsUtil.getBoolean(properties, EsResource.HTTP_SSL_ENABLED);
         }
 
+        if (properties.containsKey(EsResource.LIKE_PUSH_DOWN)) {
+            likePushDown = EsUtil.getBoolean(properties, EsResource.LIKE_PUSH_DOWN);
+        }
+
         if (StringUtils.isNotBlank(properties.get(EsResource.TYPE))) {
             mappingType = properties.get(EsResource.TYPE).trim();
         }
@@ -197,6 +206,7 @@ public class EsTable extends Table {
         tableContext.put("maxDocValueFields", String.valueOf(maxDocValueFields));
         tableContext.put(EsResource.NODES_DISCOVERY, String.valueOf(nodesDiscovery));
         tableContext.put(EsResource.HTTP_SSL_ENABLED, String.valueOf(httpSslEnabled));
+        tableContext.put(EsResource.LIKE_PUSH_DOWN, String.valueOf(likePushDown));
     }
 
     @Override
@@ -275,6 +285,8 @@ public class EsTable extends Table {
                 EsResource.NODES_DISCOVERY_DEFAULT_VALUE));
         httpSslEnabled = Boolean.parseBoolean(tableContext.getOrDefault(EsResource.HTTP_SSL_ENABLED,
                 EsResource.HTTP_SSL_ENABLED_DEFAULT_VALUE));
+        likePushDown = Boolean.parseBoolean(tableContext.getOrDefault(EsResource.LIKE_PUSH_DOWN,
+                EsResource.LIKE_PUSH_DOWN_DEFAULT_VALUE));
         PartitionType partType = PartitionType.valueOf(Text.readString(in));
         if (partType == PartitionType.UNPARTITIONED) {
             partitionInfo = SinglePartitionInfo.read(in);
