+++ fe/fe-core/src/main/java/org/apache/doris/catalog/external/EsExternalTable.java
@@ -96,6 +96,7 @@ public class EsExternalTable extends ExternalTable {
         esTable.setEnableKeywordSniff(esCatalog.enableKeywordSniff());
         esTable.setNodesDiscovery(esCatalog.enableNodesDiscovery());
         esTable.setHttpSslEnabled(esCatalog.enableSsl());
+        esTable.setLikePushDown(esCatalog.enableLikePushdown());
         esTable.setSeeds(esCatalog.getNodes());
         esTable.setHosts(String.join(",", esCatalog.getNodes()));
         esTable.syncTableMetaData();
