+++ fe/fe-core/src/main/java/org/apache/doris/datasource/EsExternalCatalog.java
@@ -106,6 +106,11 @@ public class EsExternalCatalog extends ExternalCatalog {
                 EsResource.MAPPING_ES_ID_DEFAULT_VALUE));
     }
 
+    public boolean enableLikePushdown() {
+        return Boolean.parseBoolean(catalogProperty.getOrDefault(EsResource.LIKE_PUSH_DOWN,
+                EsResource.LIKE_PUSH_DOWN_DEFAULT_VALUE));
+    }
+
     @Override
     protected void initLocalObjectsImpl() {
         esRestClient = new EsRestClient(getNodes(), getUsername(), getPassword(), enableSsl());
