+++ fe/fe-core/src/main/java/org/apache/doris/external/elasticsearch/EsUtil.java
@@ -213,167 +213,6 @@ public class EsUtil {
         return properties;
     }
 
-    private static QueryBuilder toCompoundEsDsl(Expr expr, List<Expr> notPushDownList,
-            Map<String, String> fieldsContext) {
-        CompoundPredicate compoundPredicate = (CompoundPredicate) expr;
-        switch (compoundPredicate.getOp()) {
-            case AND: {
-                QueryBuilder left = toEsDsl(compoundPredicate.getChild(0), notPushDownList, fieldsContext);
-                QueryBuilder right = toEsDsl(compoundPredicate.getChild(1), notPushDownList, fieldsContext);
-                if (left != null && right != null) {
-                    return QueryBuilders.boolQuery().must(left).must(right);
-                }
-                return null;
-            }
-            case OR: {
-                int beforeSize = notPushDownList.size();
-                QueryBuilder left = toEsDsl(compoundPredicate.getChild(0), notPushDownList, fieldsContext);
-                QueryBuilder right = toEsDsl(compoundPredicate.getChild(1), notPushDownList, fieldsContext);
-                int afterSize = notPushDownList.size();
-                if (left != null && right != null) {
-                    return QueryBuilders.boolQuery().should(left).should(right);
-                }
-                // One 'or' association cannot be pushed down and the other cannot be pushed down
-                if (afterSize > beforeSize) {
-                    if (left != null) {
-                        // add right if right don't pushdown
-                        notPushDownList.add(compoundPredicate.getChild(0));
-                    } else if (right != null) {
-                        // add left if left don't pushdown
-                        notPushDownList.add(compoundPredicate.getChild(1));
-                    }
-                }
-                return null;
-            }
-            case NOT: {
-                QueryBuilder child = toEsDsl(compoundPredicate.getChild(0), notPushDownList, fieldsContext);
-                if (child != null) {
-                    return QueryBuilders.boolQuery().mustNot(child);
-                }
-                return null;
-            }
-            default:
-                return null;
-        }
-    }
-
-    private static Expr exprWithoutCast(Expr expr) {
-        if (expr instanceof CastExpr) {
-            return exprWithoutCast(expr.getChild(0));
-        }
-        return expr;
-    }
-
-    public static QueryBuilder toEsDsl(Expr expr) {
-        return toEsDsl(expr, new ArrayList<>(), new HashMap<>());
-    }
-
-    /**
-     * Doris expr to es dsl.
-     **/
-    public static QueryBuilder toEsDsl(Expr expr, List<Expr> notPushDownList, Map<String, String> fieldsContext) {
-        if (expr == null) {
-            return null;
-        }
-        // CompoundPredicate, `between` also converted to CompoundPredicate.
-        if (expr instanceof CompoundPredicate) {
-            return toCompoundEsDsl(expr, notPushDownList, fieldsContext);
-        }
-        TExprOpcode opCode = expr.getOpcode();
-        String column;
-        Expr leftExpr = expr.getChild(0);
-        // Type transformed cast can not pushdown
-        if (leftExpr instanceof CastExpr) {
-            Expr withoutCastExpr = exprWithoutCast(leftExpr);
-            // pushdown col(float) >= 3
-            if (withoutCastExpr.getType().equals(leftExpr.getType()) || (withoutCastExpr.getType().isFloatingPointType()
-                    && leftExpr.getType().isFloatingPointType())) {
-                column = ((SlotRef) withoutCastExpr).getColumnName();
-            } else {
-                notPushDownList.add(expr);
-                return null;
-            }
-        } else if (leftExpr instanceof SlotRef) {
-            column = ((SlotRef) leftExpr).getColumnName();
-        } else {
-            notPushDownList.add(expr);
-            return null;
-        }
-        // Replace col with col.keyword if mapping exist.
-        column = fieldsContext.getOrDefault(column, column);
-        if (expr instanceof BinaryPredicate) {
-            Object value = toDorisLiteral(expr.getChild(1));
-            switch (opCode) {
-                case EQ:
-                case EQ_FOR_NULL:
-                    return QueryBuilders.termQuery(column, value);
-                case NE:
-                    return QueryBuilders.boolQuery().mustNot(QueryBuilders.termQuery(column, value));
-                case GE:
-                    return QueryBuilders.rangeQuery(column).gte(value);
-                case GT:
-                    return QueryBuilders.rangeQuery(column).gt(value);
-                case LE:
-                    return QueryBuilders.rangeQuery(column).lte(value);
-                case LT:
-                    return QueryBuilders.rangeQuery(column).lt(value);
-                default:
-                    return null;
-            }
-        }
-        if (expr instanceof IsNullPredicate) {
-            IsNullPredicate isNullPredicate = (IsNullPredicate) expr;
-            if (isNullPredicate.isNotNull()) {
-                return QueryBuilders.existsQuery(column);
-            }
-            return QueryBuilders.boolQuery().mustNot(QueryBuilders.existsQuery(column));
-        }
-        if (expr instanceof LikePredicate) {
-            LikePredicate likePredicate = (LikePredicate) expr;
-            if (likePredicate.getOp().equals(Operator.LIKE)) {
-                char[] chars = likePredicate.getChild(1).getStringValue().toCharArray();
-                // example of translation :
-                //      abc_123  ===> abc?123
-                //      abc%ykz  ===> abc*123
-                //      %abc123  ===> *abc123
-                //      _abc123  ===> ?abc123
-                //      \\_abc1  ===> \\_abc1
-                //      abc\\_123 ===> abc\\_123
-                //      abc\\%123 ===> abc\\%123
-                // NOTE. user must input sql like 'abc\\_123' or 'abc\\%ykz'
-                for (int i = 0; i < chars.length; i++) {
-                    if (chars[i] == '_' || chars[i] == '%') {
-                        if (i == 0) {
-                            chars[i] = (chars[i] == '_') ? '?' : '*';
-                        } else if (chars[i - 1] != '\\') {
-                            chars[i] = (chars[i] == '_') ? '?' : '*';
-                        }
-                    }
-                }
-                return QueryBuilders.wildcardQuery(column, new String(chars));
-            } else {
-                return QueryBuilders.wildcardQuery(column, likePredicate.getChild(1).getStringValue());
-            }
-        }
-        if (expr instanceof InPredicate) {
-            InPredicate inPredicate = (InPredicate) expr;
-            List<Object> values = inPredicate.getListChildren().stream().map(EsUtil::toDorisLiteral)
-                    .collect(Collectors.toList());
-            if (inPredicate.isNotIn()) {
-                return QueryBuilders.boolQuery().mustNot(QueryBuilders.termsQuery(column, values));
-            }
-            return QueryBuilders.termsQuery(column, values);
-        }
-        if (expr instanceof FunctionCallExpr) {
-            FunctionCallExpr functionCallExpr = (FunctionCallExpr) expr;
-            if ("esquery".equals(functionCallExpr.getFnName().getFunction())) {
-                String stringValue = functionCallExpr.getChild(1).getStringValue();
-                return new QueryBuilders.EsQueryBuilder(stringValue);
-            }
-        }
-        return null;
-    }
-
 
     /**
      * Generate columns from ES Cluster.
@@ -456,29 +295,6 @@ public class EsUtil {
             default:
                 return Type.UNSUPPORTED;
         }
-    }
-
-    private static Object toDorisLiteral(Expr expr) {
-        if (!expr.isLiteral()) {
-            return null;
-        }
-        if (expr instanceof BoolLiteral) {
-            BoolLiteral boolLiteral = (BoolLiteral) expr;
-            return boolLiteral.getValue();
-        } else if (expr instanceof DecimalLiteral) {
-            DecimalLiteral decimalLiteral = (DecimalLiteral) expr;
-            return decimalLiteral.getValue();
-        } else if (expr instanceof FloatLiteral) {
-            FloatLiteral floatLiteral = (FloatLiteral) expr;
-            return floatLiteral.getValue();
-        } else if (expr instanceof IntLiteral) {
-            IntLiteral intLiteral = (IntLiteral) expr;
-            return intLiteral.getValue();
-        } else if (expr instanceof LargeIntLiteral) {
-            LargeIntLiteral largeIntLiteral = (LargeIntLiteral) expr;
-            return largeIntLiteral.getLongValue();
-        }
-        return expr.getStringValue();
     }
 
 }
