+++ fe/fe-core/src/main/java/org/apache/doris/external/elasticsearch/QueryBuilders.java
@@ -17,6 +17,24 @@
 
 package org.apache.doris.external.elasticsearch;
 
+import org.apache.doris.analysis.BinaryPredicate;
+import org.apache.doris.analysis.BoolLiteral;
+import org.apache.doris.analysis.CastExpr;
+import org.apache.doris.analysis.CompoundPredicate;
+import org.apache.doris.analysis.DecimalLiteral;
+import org.apache.doris.analysis.Expr;
+import org.apache.doris.analysis.FloatLiteral;
+import org.apache.doris.analysis.FunctionCallExpr;
+import org.apache.doris.analysis.InPredicate;
+import org.apache.doris.analysis.IntLiteral;
+import org.apache.doris.analysis.IsNullPredicate;
+import org.apache.doris.analysis.LargeIntLiteral;
+import org.apache.doris.analysis.LikePredicate;
+import org.apache.doris.analysis.LikePredicate.Operator;
+import org.apache.doris.analysis.SlotRef;
+import org.apache.doris.catalog.EsResource;
+import org.apache.doris.thrift.TExprOpcode;
+
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -26,10 +44,13 @@ import org.apache.logging.log4j.Logger;
 import java.io.IOException;
 import java.io.StringWriter;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
+import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Objects;
+import java.util.stream.Collectors;
 
 
 /**
@@ -37,6 +58,238 @@ import java.util.Objects;
  * Some query builders and static helper method have been copied from Elasticsearch
  */
 public final class QueryBuilders {
+
+    /**
+     * Generate dsl from compound expr.
+     **/
+    private static QueryBuilder toCompoundEsDsl(Expr expr, List<Expr> notPushDownList,
+            Map<String, String> fieldsContext, boolean likePushdown) {
+        CompoundPredicate compoundPredicate = (CompoundPredicate) expr;
+        switch (compoundPredicate.getOp()) {
+            case AND: {
+                QueryBuilder left = toEsDsl(compoundPredicate.getChild(0), notPushDownList, fieldsContext,
+                        likePushdown);
+                QueryBuilder right = toEsDsl(compoundPredicate.getChild(1), notPushDownList, fieldsContext,
+                        likePushdown);
+                if (left != null && right != null) {
+                    return QueryBuilders.boolQuery().must(left).must(right);
+                }
+                return null;
+            }
+            case OR: {
+                int beforeSize = notPushDownList.size();
+                QueryBuilder left = toEsDsl(compoundPredicate.getChild(0), notPushDownList, fieldsContext,
+                        likePushdown);
+                QueryBuilder right = toEsDsl(compoundPredicate.getChild(1), notPushDownList, fieldsContext,
+                        likePushdown);
+                int afterSize = notPushDownList.size();
+                if (left != null && right != null) {
+                    return QueryBuilders.boolQuery().should(left).should(right);
+                }
+                // One 'or' association cannot be pushed down and the other cannot be pushed down
+                if (afterSize > beforeSize) {
+                    if (left != null) {
+                        // add right if right don't pushdown
+                        notPushDownList.add(compoundPredicate.getChild(0));
+                    } else if (right != null) {
+                        // add left if left don't pushdown
+                        notPushDownList.add(compoundPredicate.getChild(1));
+                    }
+                }
+                return null;
+            }
+            case NOT: {
+                QueryBuilder child = toEsDsl(compoundPredicate.getChild(0), notPushDownList, fieldsContext,
+                        likePushdown);
+                if (child != null) {
+                    return QueryBuilders.boolQuery().mustNot(child);
+                }
+                return null;
+            }
+            default:
+                return null;
+        }
+    }
+
+    /**
+     * Get the expr inside the cast.
+     **/
+    private static Expr exprWithoutCast(Expr expr) {
+        if (expr instanceof CastExpr) {
+            return exprWithoutCast(expr.getChild(0));
+        }
+        return expr;
+    }
+
+    public static QueryBuilder toEsDsl(Expr expr) {
+        return toEsDsl(expr, new ArrayList<>(), new HashMap<>(),
+                Boolean.parseBoolean(EsResource.LIKE_PUSH_DOWN_DEFAULT_VALUE));
+    }
+
+    private static QueryBuilder parseBinaryPredicate(Expr expr, TExprOpcode opCode, String column) {
+        Object value = toDorisLiteral(expr.getChild(1));
+        switch (opCode) {
+            case EQ:
+            case EQ_FOR_NULL:
+                return QueryBuilders.termQuery(column, value);
+            case NE:
+                return QueryBuilders.boolQuery().mustNot(QueryBuilders.termQuery(column, value));
+            case GE:
+                return QueryBuilders.rangeQuery(column).gte(value);
+            case GT:
+                return QueryBuilders.rangeQuery(column).gt(value);
+            case LE:
+                return QueryBuilders.rangeQuery(column).lte(value);
+            case LT:
+                return QueryBuilders.rangeQuery(column).lt(value);
+            default:
+                return null;
+        }
+    }
+
+    private static QueryBuilder parseIsNullPredicate(Expr expr, String column) {
+        IsNullPredicate isNullPredicate = (IsNullPredicate) expr;
+        if (isNullPredicate.isNotNull()) {
+            return QueryBuilders.existsQuery(column);
+        }
+        return QueryBuilders.boolQuery().mustNot(QueryBuilders.existsQuery(column));
+    }
+
+    private static QueryBuilder parseLikePredicate(Expr expr, String column) {
+        LikePredicate likePredicate = (LikePredicate) expr;
+        if (likePredicate.getOp().equals(Operator.LIKE)) {
+            char[] chars = likePredicate.getChild(1).getStringValue().toCharArray();
+            // example of translation :
+            //      abc_123  ===> abc?123
+            //      abc%ykz  ===> abc*123
+            //      %abc123  ===> *abc123
+            //      _abc123  ===> ?abc123
+            //      \\_abc1  ===> \\_abc1
+            //      abc\\_123 ===> abc\\_123
+            //      abc\\%123 ===> abc\\%123
+            // NOTE. user must input sql like 'abc\\_123' or 'abc\\%ykz'
+            for (int i = 0; i < chars.length; i++) {
+                if (chars[i] == '_' || chars[i] == '%') {
+                    if (i == 0) {
+                        chars[i] = (chars[i] == '_') ? '?' : '*';
+                    } else if (chars[i - 1] != '\\') {
+                        chars[i] = (chars[i] == '_') ? '?' : '*';
+                    }
+                }
+            }
+            return QueryBuilders.wildcardQuery(column, new String(chars));
+        } else {
+            return QueryBuilders.wildcardQuery(column, likePredicate.getChild(1).getStringValue());
+        }
+    }
+
+    private static QueryBuilder parseInPredicate(Expr expr, String column) {
+        InPredicate inPredicate = (InPredicate) expr;
+        List<Object> values = inPredicate.getListChildren().stream().map(QueryBuilders::toDorisLiteral)
+                .collect(Collectors.toList());
+        if (inPredicate.isNotIn()) {
+            return QueryBuilders.boolQuery().mustNot(QueryBuilders.termsQuery(column, values));
+        }
+        return QueryBuilders.termsQuery(column, values);
+    }
+
+    private static QueryBuilder parseFunctionCallExpr(Expr expr) {
+        // esquery(k1, '{
+        //        "match_phrase": {
+        //           "k1": "doris on es"
+        //        }
+        //    }');
+        // The first child k1 compatible with expr syntax
+        FunctionCallExpr functionCallExpr = (FunctionCallExpr) expr;
+        if ("esquery".equals(functionCallExpr.getFnName().getFunction())) {
+            String stringValue = functionCallExpr.getChild(1).getStringValue();
+            return new QueryBuilders.EsQueryBuilder(stringValue);
+        }
+        return null;
+    }
+
+    /**
+     * Doris expr to es dsl.
+     **/
+    public static QueryBuilder toEsDsl(Expr expr, List<Expr> notPushDownList, Map<String, String> fieldsContext,
+            boolean likePushDown) {
+        if (expr == null) {
+            return null;
+        }
+        // CompoundPredicate, `between` also converted to CompoundPredicate.
+        if (expr instanceof CompoundPredicate) {
+            return toCompoundEsDsl(expr, notPushDownList, fieldsContext, likePushDown);
+        }
+        TExprOpcode opCode = expr.getOpcode();
+        String column;
+        Expr leftExpr = expr.getChild(0);
+        // Type transformed cast can not pushdown
+        if (leftExpr instanceof CastExpr) {
+            Expr withoutCastExpr = exprWithoutCast(leftExpr);
+            // pushdown col(float) >= 3
+            if (withoutCastExpr.getType().equals(leftExpr.getType()) || (withoutCastExpr.getType().isFloatingPointType()
+                    && leftExpr.getType().isFloatingPointType())) {
+                column = ((SlotRef) withoutCastExpr).getColumnName();
+            } else {
+                notPushDownList.add(expr);
+                return null;
+            }
+        } else if (leftExpr instanceof SlotRef) {
+            column = ((SlotRef) leftExpr).getColumnName();
+        } else {
+            notPushDownList.add(expr);
+            return null;
+        }
+        // Replace col with col.keyword if mapping exist.
+        column = fieldsContext.getOrDefault(column, column);
+        if (expr instanceof BinaryPredicate) {
+            return parseBinaryPredicate(expr, opCode, column);
+        }
+        if (expr instanceof IsNullPredicate) {
+            return parseIsNullPredicate(expr, column);
+        }
+        if (expr instanceof LikePredicate) {
+            if (!likePushDown) {
+                notPushDownList.add(expr);
+                return null;
+            } else {
+                return parseLikePredicate(expr, column);
+            }
+        }
+        if (expr instanceof InPredicate) {
+            return parseInPredicate(expr, column);
+        }
+        if (expr instanceof FunctionCallExpr) {
+            return parseFunctionCallExpr(expr);
+        }
+        return null;
+    }
+
+    /**
+     * Expr trans to doris literal.
+     **/
+    private static Object toDorisLiteral(Expr expr) {
+        if (!expr.isLiteral()) {
+            return null;
+        }
+        if (expr instanceof BoolLiteral) {
+            BoolLiteral boolLiteral = (BoolLiteral) expr;
+            return boolLiteral.getValue();
+        } else if (expr instanceof DecimalLiteral) {
+            DecimalLiteral decimalLiteral = (DecimalLiteral) expr;
+            return decimalLiteral.getValue();
+        } else if (expr instanceof FloatLiteral) {
+            FloatLiteral floatLiteral = (FloatLiteral) expr;
+            return floatLiteral.getValue();
+        } else if (expr instanceof IntLiteral) {
+            IntLiteral intLiteral = (IntLiteral) expr;
+            return intLiteral.getValue();
+        } else if (expr instanceof LargeIntLiteral) {
+            LargeIntLiteral largeIntLiteral = (LargeIntLiteral) expr;
+            return largeIntLiteral.getLongValue();
+        }
+        return expr.getStringValue();
+    }
 
     /**
      * A query that matches on all documents.
@@ -48,7 +301,7 @@ public final class QueryBuilders {
     /**
      * A Query that matches documents containing a term.
      *
-     * @param name  The name of the field
+     * @param name The name of the field
      * @param value The value of the term
      */
     public static TermQueryBuilder termQuery(String name, String value) {
@@ -58,7 +311,7 @@ public final class QueryBuilders {
     /**
      * A Query that matches documents containing a term.
      *
-     * @param name  The name of the field
+     * @param name The name of the field
      * @param value The value of the term
      */
     public static TermQueryBuilder termQuery(String name, int value) {
@@ -68,7 +321,7 @@ public final class QueryBuilders {
     /**
      * A Query that matches documents containing a term.
      *
-     * @param name  The name of the field
+     * @param name The name of the field
      * @param value The value of the term
      */
     public static TermQueryBuilder termQuery(String name, long value) {
@@ -78,7 +331,7 @@ public final class QueryBuilders {
     /**
      * A Query that matches documents containing a term.
      *
-     * @param name  The name of the field
+     * @param name The name of the field
      * @param value The value of the term
      */
     public static TermQueryBuilder termQuery(String name, float value) {
@@ -88,7 +341,7 @@ public final class QueryBuilders {
     /**
      * A Query that matches documents containing a term.
      *
-     * @param name  The name of the field
+     * @param name The name of the field
      * @param value The value of the term
      */
     public static TermQueryBuilder termQuery(String name, double value) {
@@ -98,7 +351,7 @@ public final class QueryBuilders {
     /**
      * A Query that matches documents containing a term.
      *
-     * @param name  The name of the field
+     * @param name The name of the field
      * @param value The value of the term
      */
     public static TermQueryBuilder termQuery(String name, boolean value) {
@@ -108,7 +361,7 @@ public final class QueryBuilders {
     /**
      * A Query that matches documents containing a term.
      *
-     * @param name  The name of the field
+     * @param name The name of the field
      * @param value The value of the term
      */
     public static TermQueryBuilder termQuery(String name, Object value) {
@@ -123,7 +376,7 @@ public final class QueryBuilders {
      * a Wildcard term should not start with one of the wildcards {@code *} or
      * {@code ?}.
      *
-     * @param name  The field name
+     * @param name The field name
      * @param query The wildcard query string
      */
     public static WildcardQueryBuilder wildcardQuery(String name, String query) {
@@ -141,7 +394,7 @@ public final class QueryBuilders {
     /**
      * A filter for a field based on several terms matching on any of them.
      *
-     * @param name   The field name
+     * @param name The field name
      * @param values The terms
      */
     public static TermsQueryBuilder termsQuery(String name, Iterable<?> values) {
@@ -505,7 +758,7 @@ public final class QueryBuilders {
     /**
      * Write (scalar) value (string, number, boolean or null) to json format
      *
-     * @param out   source target
+     * @param out source target
      * @param value value to write
      * @throws IOException if error
      */
